requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiogram-media-group==0.5.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.2
attrs==25.3.0
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
distro==1.9.0
et_xmlfile==2.0.0
fonttools==4.61.1
frozenlist==1.6.0
geojson==3.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.64.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
kiwisolver==1.4.9
magic-filter==1.0.12
matplotlib==3.10.8
multidict==6.4.3
numpy==2.3.4
openpyxl==3.1.5
openpyxl-utils==0.0.3
packaging==25.0
pandas==2.3.3
pillow==12.1.1
pluggy==1.6.0
propcache==0.3.1
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.11.4
pydantic_core==2.33.2
pygame==2.6.1
Pygments==2.19.2
pyowm==3.5.0
pyparsing==3.3.2
PySocks==1.7.1
pyTelegramBotAPI==4.28.0
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==20.3
pytz==2025.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
telebot==0.0.5
telegram==0.0.1
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
utils==1.0.2
websockets==15.0.1
wheel==0.45.1
yarl==1.20.0
